+++ src/components/about/about.config.js
@@ -1,4 +1,4 @@
-import aboutImage from '../../assets/icons/placeholder/undraw_profile_details_re_ch9r.svg';
+import aboutSVG from '../../assets/icons/placeholder/undraw_profile_details_re_ch9r.svg';
 import gitHubIcon from '../../assets/icons/github_mark/github-mark-white.svg';
 import linkedinIcon from '../../assets/icons/linkedin_mark/linkedin_00.svg';
 import xIcon from '../../assets/icons/x_mark/x_01.svg';
@@ -24,7 +24,7 @@ export default {
             {
               element: 'img',
               attributes: {
-                src: aboutImage,
+                src: aboutSVG,
               },
             },
             {
@@ -72,6 +72,7 @@ export default {
                         {
                           element: 'img',
                           attributes: {
+                            class: 'icon icon_social',
                             src: gitHubIcon,
                             onload: 'SVGInject(this)',
                           },
@@ -92,6 +93,7 @@ export default {
                         {
                           element: 'img',
                           attributes: {
+                            class: 'icon icon_social',
                             src: linkedinIcon,
                             onload: 'SVGInject(this)',
                           },
@@ -112,6 +114,7 @@ export default {
                         {
                           element: 'img',
                           attributes: {
+                            class: 'icon icon_social',
                             src: xIcon,
                             onload: 'SVGInject(this)',
                           },