+++ src/components/projects/projects.config.js
@@ -1,3 +1,6 @@
+import gitHubIcon from '../../assets/icons/github_mark/github-mark-white.svg';
+import openInNewIcon from '../../assets/icons/open_in_new.svg';
+
 export default {
   element: 'section',
   attributes: {
@@ -5,7 +8,456 @@ export default {
   },
   children: [
     {
+      element: 'h2',
+      attributes: {
+        textContent: 'Projects',
+      },
+    },
+    {
       element: 'article',
+      children: [
+        {
+          element: 'picture',
+          children: [
+            {
+              element: 'img',
+              attributes: {
+                src: 'https://placehold.co/600x400',
+              },
+            },
+          ],
+        },
+        {
+          element: 'div',
+          attributes: {
+            class: 'nameLater',
+          },
+          children: [
+            {
+              element: 'h3',
+              attributes: {
+                textContent: 'Project00',
+              },
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: gitHubIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: openInNewIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+          ],
+        },
+        {
+          element: 'p',
+          attributes: {
+            textContent:
+              'Maecenas fermentum purus diam, a gravida nunc mattis ac. Nulla volutpat lectus odio, in consectetur felis ultricies id. Nulla ex justo, sollicitudin ac faucibus eu, porttitor nec tortor. Nunc sit amet sapien ex.',
+          },
+        },
+      ],
+    },
+    {
+      element: 'article',
+      children: [
+        {
+          element: 'picture',
+          children: [
+            {
+              element: 'img',
+              attributes: {
+                src: 'https://placehold.co/600x400',
+              },
+            },
+          ],
+        },
+        {
+          element: 'div',
+          attributes: {
+            class: 'nameLater',
+          },
+          children: [
+            {
+              element: 'h3',
+              attributes: {
+                textContent: 'Project01',
+              },
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: gitHubIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: openInNewIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+          ],
+        },
+        {
+          element: 'p',
+          attributes: {
+            textContent:
+              'Nulla scelerisque felis vel nibh sagittis sollicitudin. Morbi urna lacus, ullamcorper et aliquet et, mollis vestibulum elit. Sed vitae ornare lectus. Nam quis mattis metus.',
+          },
+        },
+      ],
+    },
+    {
+      element: 'article',
+      children: [
+        {
+          element: 'picture',
+          children: [
+            {
+              element: 'img',
+              attributes: {
+                src: 'https://placehold.co/600x400',
+              },
+            },
+          ],
+        },
+        {
+          element: 'div',
+          attributes: {
+            class: 'nameLater',
+          },
+          children: [
+            {
+              element: 'h3',
+              attributes: {
+                textContent: 'Project02',
+              },
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: gitHubIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: openInNewIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: gitHubIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: openInNewIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+          ],
+        },
+        {
+          element: 'p',
+          attributes: {
+            textContent:
+              'Donec ac ullamcorper odio, tristique laoreet tellus. Nulla tempor enim lorem, ac iaculis urna pulvinar non. Ut finibus nisi mauris. Integer dignissim nisi nec sapien cursus posuere. Ut et ipsum nisi. In vel elit nulla.',
+          },
+        },
+      ],
+    },
+    {
+      element: 'article',
+      children: [
+        {
+          element: 'picture',
+          children: [
+            {
+              element: 'img',
+              attributes: {
+                src: 'https://placehold.co/600x400',
+              },
+            },
+          ],
+        },
+        {
+          element: 'div',
+          attributes: {
+            class: 'nameLater',
+          },
+          children: [
+            {
+              element: 'h3',
+              attributes: {
+                textContent: 'Project03',
+              },
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: gitHubIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: openInNewIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+          ],
+        },
+        {
+          element: 'p',
+          attributes: {
+            textContent:
+              'Suspendisse ornare rhoncus tincidunt. Suspendisse vulputate aliquam eros in blandit. Pellentesque sed tempor dolor. Donec eleifend, augue sit amet tincidunt ornare, lacus lorem malesuada metus, nec suscipit dolor tortor non eros.',
+          },
+        },
+      ],
+    },
+    {
+      element: 'article',
+      children: [
+        {
+          element: 'picture',
+          children: [
+            {
+              element: 'img',
+              attributes: {
+                src: 'https://placehold.co/600x400',
+              },
+            },
+          ],
+        },
+        {
+          element: 'div',
+          attributes: {
+            class: 'nameLater',
+          },
+          children: [
+            {
+              element: 'h3',
+              attributes: {
+                textContent: 'Project04',
+              },
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: gitHubIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: openInNewIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+          ],
+        },
+        {
+          element: 'p',
+          attributes: {
+            textContent:
+              'Interdum et malesuada fames ac ante ipsum primis in faucibus. Curabitur faucibus magna quis magna blandit, sed ultricies felis ornare. Maecenas vestibulum ac magna nec fermentum.',
+          },
+        },
+      ],
+    },
+    {
+      element: 'article',
+      children: [
+        {
+          element: 'picture',
+          children: [
+            {
+              element: 'img',
+              attributes: {
+                src: 'https://placehold.co/600x400',
+              },
+            },
+          ],
+        },
+        {
+          element: 'div',
+          attributes: {
+            class: 'nameLater',
+          },
+          children: [
+            {
+              element: 'h3',
+              attributes: {
+                textContent: 'Project05',
+              },
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: gitHubIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+            {
+              element: 'a',
+              attributes: {
+                href: '#',
+              },
+              children: [
+                {
+                  element: 'img',
+                  attributes: {
+                    class: 'icon icon_project',
+                    src: openInNewIcon,
+                    onload: 'SVGInject(this)',
+                  },
+                },
+              ],
+            },
+          ],
+        },
+        {
+          element: 'p',
+          attributes: {
+            textContent:
+              'Suspendisse commodo pharetra sem, euismod suscipit velit condimentum in. Integer rutrum malesuada orci, id euismod augue ultrices sed. Nulla eu enim ultricies, tincidunt lorem at, efficitur ex. Aenean tempus risus ac purus aliquet finibus.',
+          },
+        },
+      ],
     },
   ],
 };