+++ src/helpers/pubSub.js
@@ -6,7 +6,7 @@ export default {
       this.subscribers[subscriber].push(fn);
     }
   },
-  unsubsccribe(subscriber, fn) {
+  unsubscribe(subscriber, fn) {
     if (this.subscribers[subscriber]) {
       this.subscribers[subscriber].splice(this.subscribers[subscriber].indexOf(fn), 1);
       if (this.subscribers[subscriber].length === 0) delete this.subscribers[subscriber];